+++ @types/function/chat_message.d.ts
@@ -126,7 +126,7 @@ interface SetChatMessagesOption {
  * // 补充倒数第二楼的楼层变量
  * const chat_message = getChatMessages(-2)[0];
  * _.set(chat_message.data, '神乐光好感度', 5);
- * await setChatMessages([{message_id: 0, data: chat_message.data}]);
+ * await setChatMessages([{message_id: 0, data: chat_message.data}], {refresh: 'none'});
  *
  * @example
  * // 切换为开局 3